cleanups of commented code and adjusting import shortcuts
